Reformat and move revng python modules
