+++ python/revng/merge_dynamic/__init__.py
