+++ scripts/revng-dump-model
@@ -2,7 +2,7 @@
 
 import sys
 
-from revng_dump_model.__main__ import main
+from revng.revng_dump_model.__main__ import main
 
 if __name__ == "__main__":
     sys.exit(main())
