+++ scripts/revng-merge-dynamic
@@ -1,6 +1,6 @@
 #!/usr/bin/env python3
 
-from revng_merge_dynamic.__main__ import main
+from revng.merge_dynamic.__main__ import main
 
 if __name__ == "__main__":
     main()
