+++ tests/runtime/RuntimeTests.cmake
@@ -8,7 +8,11 @@ macro(artifact_handler CATEGORY INPUT_FILE CONFIGURATION OUTPUT TARGET_NAME)
   list(GET INPUT_FILE 1 COMPILED_RUN_INPUT)
 
   if("${CATEGORY}" STREQUAL "tests_runtime" AND NOT "${CONFIGURATION}" STREQUAL "static_native")
-    set(COMMAND_TO_RUN "./bin/revng" translate -i ${COMPILED_INPUT} -o "${OUTPUT}")
+    set(
+      COMMAND_TO_RUN
+      ${CMAKE_COMMAND} -E env "PYTHONPATH=${CMAKE_BINARY_DIR}/python"
+      "./bin/revng" translate -i ${COMPILED_INPUT} -o "${OUTPUT}"
+    )
     set(DEPEND_ON revng-all-binaries)
 
     if(NOT "${CONFIGURATION}" STREQUAL "aarch64")
@@ -35,6 +39,7 @@ macro(artifact_handler CATEGORY INPUT_FILE CONFIGURATION OUTPUT TARGET_NAME)
 
   if("${CATEGORY}" STREQUAL "tests_runtime" AND NOT "${CONFIGURATION}" STREQUAL "static_native")
     set(COMMAND_TO_RUN
+      "${CMAKE_COMMAND}" -E env "PYTHONPATH=${CMAKE_BINARY_DIR}/python"
       "./bin/revng"
       lift
       ${COMPILED_INPUT}
@@ -47,6 +52,7 @@ register_derived_artifact("compiled;compiled-run" "lifted" ".ll" "FILE")
 macro(artifact_handler CATEGORY INPUT_FILE CONFIGURATION OUTPUT TARGET_NAME)
   if("${CATEGORY}" STREQUAL "tests_runtime" AND NOT "${CONFIGURATION}" STREQUAL "static_native")
     set(COMMAND_TO_RUN
+      "${CMAKE_COMMAND}" -E env "PYTHONPATH=${CMAKE_BINARY_DIR}/python"
       "./bin/revng"
       opt
       "${INPUT_FILE}"
@@ -67,6 +73,7 @@ register_derived_artifact("lifted" "abi-enforced-for-decompilation" ".bc" "FILE"
 macro(artifact_handler CATEGORY INPUT_FILE CONFIGURATION OUTPUT TARGET_NAME)
   if("${CATEGORY}" STREQUAL "tests_runtime" AND NOT "${CONFIGURATION}" STREQUAL "static_native" AND "${TARGET_NAME}" MATCHES "calc")
     set(COMMAND_TO_RUN
+      "${CMAKE_COMMAND}" -E env "PYTHONPATH=${CMAKE_BINARY_DIR}/python"
       "${CMAKE_CURRENT_BINARY_DIR}/bin/revng"
       opt
       "${INPUT_FILE}"
